我这里int64变成int32了卧槽

+++ 学校上课写的/6.HomeWork_5/TurnDown.c
@@ -2,9 +2,9 @@
 #include <stdio.h>
 
 int main(){
-    int num, mod, backup, i = 0, weight = 1;
+    long long num, mod, backup, i = 0, weight = 1;
     printf("请输入一个非负整数:");
-    scanf("%d", &num);
+    scanf("%llu", &num);
     backup = num;
     printf("逆序为:");
     do{
@@ -12,13 +12,13 @@ int main(){
         backup /= 10;
         i++;
         weight *= 10;
-        printf("%d", mod);
+        printf("%llu", mod);
     }while(backup >= 1);
     weight /= 10;
     printf("\n它是一个%d位数", i);
     printf("\n各位数字分别为:");
     do{
-        printf("%d ", num / weight);
+        printf("%llu ", num / weight);
         num -= (num / weight) * weight;
         weight /= 10;
     }while(weight >= 1);
